+++ src/NumSharp.Core/Utilities/Converts.cs
@@ -180,7 +180,7 @@ namespace NumSharp.Utilities
                     return ((IConvertible)value).ToUInt64(CultureInfo.InvariantCulture);
                 case NPTypeCode.Single:
                     // Half doesn't implement IConvertible
-                    if (value is Half hs) return (float)(double)hs;
+                    if (value is Half hs) return (float)hs;
                     return ((IConvertible)value).ToSingle(CultureInfo.InvariantCulture);
                 case NPTypeCode.Double:
                     // Half doesn't implement IConvertible
@@ -470,22 +470,22 @@ namespace NumSharp.Utilities
                             throw new NotSupportedException();
                     }
                 case NPTypeCode.Half:
-                    // Half target type - convert source to double first, then to Half
+                    // Half target type - C# Half has direct casts from all numeric types except decimal
                     switch (InfoOf<T>.NPTypeCode)
                     {
-                        case NPTypeCode.Boolean: return (Half)(Unsafe.As<T, bool>(ref value) ? 1.0 : 0.0);
-                        case NPTypeCode.Byte:    return (Half)(double)Unsafe.As<T, byte>(ref value);
-                        case NPTypeCode.SByte:   return (Half)(double)Unsafe.As<T, sbyte>(ref value);
-                        case NPTypeCode.Int16:   return (Half)(double)Unsafe.As<T, short>(ref value);
-                        case NPTypeCode.UInt16:  return (Half)(double)Unsafe.As<T, ushort>(ref value);
-                        case NPTypeCode.Int32:   return (Half)(double)Unsafe.As<T, int>(ref value);
-                        case NPTypeCode.UInt32:  return (Half)(double)Unsafe.As<T, uint>(ref value);
-                        case NPTypeCode.Int64:   return (Half)(double)Unsafe.As<T, long>(ref value);
-                        case NPTypeCode.UInt64:  return (Half)(double)Unsafe.As<T, ulong>(ref value);
-                        case NPTypeCode.Char:    return (Half)(double)Unsafe.As<T, char>(ref value);
+                        case NPTypeCode.Boolean: return (Half)(Unsafe.As<T, bool>(ref value) ? 1 : 0);
+                        case NPTypeCode.Byte:    return (Half)Unsafe.As<T, byte>(ref value);
+                        case NPTypeCode.SByte:   return (Half)Unsafe.As<T, sbyte>(ref value);
+                        case NPTypeCode.Int16:   return (Half)Unsafe.As<T, short>(ref value);
+                        case NPTypeCode.UInt16:  return (Half)Unsafe.As<T, ushort>(ref value);
+                        case NPTypeCode.Int32:   return (Half)Unsafe.As<T, int>(ref value);
+                        case NPTypeCode.UInt32:  return (Half)Unsafe.As<T, uint>(ref value);
+                        case NPTypeCode.Int64:   return (Half)Unsafe.As<T, long>(ref value);
+                        case NPTypeCode.UInt64:  return (Half)Unsafe.As<T, ulong>(ref value);
+                        case NPTypeCode.Char:    return (Half)Unsafe.As<T, char>(ref value);
                         case NPTypeCode.Double:  return (Half)Unsafe.As<T, double>(ref value);
                         case NPTypeCode.Single:  return (Half)Unsafe.As<T, float>(ref value);
-                        case NPTypeCode.Decimal: return (Half)(double)Unsafe.As<T, decimal>(ref value);
+                        case NPTypeCode.Decimal: return (Half)Unsafe.As<T, decimal>(ref value);
                         default:
                             throw new NotSupportedException();
                     }
